+++ bin/actions/0220022-32.sh
@@ -7,6 +7,10 @@
 focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
+path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
+
+"$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
+
 if [[ "$focusApplication" =~ "chromium-browse" ]] \
 || [[ "$focusApplication" =~ "firefox" ]]; then
   clipboard=$( xsel -ob )
@@ -27,5 +31,9 @@ else
   else
     me=$SUDO_USER
     sudo -u "$me" nohup datagrip >/dev/null &
+    unset me
   fi
 fi
+
+unset focusApplication
+unset path_self