+++ bin/actions/0220022-33.sh
@@ -8,13 +8,15 @@
 focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
+path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
+
+"$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
+
 if [[ "$focusApplication" =~ "claws-mail" ]]; then
   xdotool sleep 0.1
   xdotool key Ctrl+Shift+f
 else
   if [[ "$focusApplication" =~ "gnome-terminal" ]]; then
-    xdotool sleep 0.1
-    xdotool key 0xff08  # Backspace
     xdotool key Ctrl+Shift+f
   else
     if pidof -s firefox >/dev/null; then
@@ -22,6 +24,10 @@ else
     else
       me=$SUDO_USER
       sudo -u "$me" nohup firefox >/dev/null &
+      unset me
     fi
   fi
 fi
+
+unset focusApplication
+unser path_self