+++ bin/actions/0220022-46-al-cr-sr.sh
@@ -2,6 +2,7 @@
 
 # workaround while "◢ + C" emits event code w/ trailing modifiers "al-cr-sr"
 
-path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
 "$path_self"/0220022-46.sh
+
 unset path_self