+++ bin/actions/0220022-46.sh
@@ -6,6 +6,10 @@
 focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
+path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
+
+"$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
+
 if [[ "$focusApplication" =~ "chromium-browse" ]] \
 || [[ "$focusApplication" =~ "firefox" ]]; then
   #  Chromium or Firefox is focussed: Copy last segment of URL
@@ -30,5 +34,9 @@ else
   else
     me=$SUDO_USER
     sudo -u "$me" nohup chromium-browser >/dev/null &  # Launch Chromium
+    unset me
   fi
 fi
+
+unset focusApplication
+unset path_self